+++ CassandraSharp/Instrumentation/NullInstrumentation.cs
@@ -25,7 +25,7 @@ namespace CassandraSharp.Instrumentation
         {
         }
 
-        public void ClientConnectionInfo(InstrumentationToken token, IPAddress coordinator, byte streamId)
+        public void ClientConnectionInfo(InstrumentationToken token, IPAddress coordinator, ushort streamId)
         {
         }
