+++ CassandraSharpUnitTests/Performance/PerformanceInstrumentation.cs
@@ -50,7 +50,7 @@ namespace CassandraSharpUnitTests.Performance
         {
         }
 
-        public void ClientConnectionInfo(InstrumentationToken token, IPAddress coordinator, byte streamId)
+        public void ClientConnectionInfo(InstrumentationToken token, IPAddress coordinator, ushort streamId)
         {
         }
